refactor(NumberInput): Create `useInputKeys` hook
